+++ src/index.ts
@@ -1,81 +1,99 @@
-import { createSyncer } from './sync.js'
-import { r } from 'rethinkdb-ts'
+import { connect } from './connect.js'
+import { sync } from './sync.js'
+import { r as _r } from 'rethinkdb-ts'
+import type { ConnectOptions } from './connect.js';
 import type { ExtraTableConfigIndexBase } from './indexes.js'
-import type { ExtraTableConfig, ExtraTableConfigTypeBase, RTableExtra } from './types.js'
-import type { Empty } from './util.js'
+import type { SyncOptions } from './sync.js'
+import type { ExtraTableConfig, ExtraTableConfigTypeBase, RDatabaseExtraConfigs, RExtra } from './types.js'
+import type { R } from 'rethinkdb-ts'
+
+export { isCursor, isRethinkDBError } from 'rethinkdb-ts'
+export type * from 'rethinkdb-ts/lib/types.js'
 
 /**
- * Constructs an object that is used in `attachConfigurations`, and does so with some nice intellisence autocomplete.
- *
- * It needs to be a chained function so that the user can specify a `Type` generic, but the generic that produces a `const` type for `indexes` internally can still be inferred.
- *
- * The different types of secondary indexes that can be configured are as follows:
- *
+ * Describe configuration for a table in a database.
+ * 
+ * Use the generic `T` to pass the type of data contained in the table. It must have an `id` property with a valid primary key type.
+ * 
+ * No need to pass any arguments, `_options` is a future placeholder - this will return another function for configuring indexes, so immediately invoke that like so:
+ * 
  * ```ts
- * simple: {} // (on 'simple' field)
- * multi: { multi: true } // (on 'multi' field)
- * compound: { compound: ['last_name', 'first_name'] }
- * custom: { custom: function (row) { return row('hobbies').add(row('sports')) } }
+ * table<YourDataType>()({ ... })
  * ```
- *
- * The `multi` index here is really just a simple index with the multi index option. Any simple or custom index can be set to be a multi index as long as the value in question is an array. A 'multi compound' index can also be constructed, but needs to be done with a custom expression. Read more about rethinkdb [secondary indexes here](https://rethinkdb.com/docs/secondary-indexes/javascript).
- *
- * @example
- * interface MyObject {
- *   id: string,
- *   my_number: boolean
- * }
- *
- * // Configure the 'myobjects' table in 'mydatabase' as holding MyObject documents,
- * // and with a custom index based on the `my_number` field
- * config<MyObject>('mydatabase', 'myobjects')(
- *   { my_number: { custom: function (row) { return row('my_number').add(1) } } }
- * )
  */
-export function configure<T extends ExtraTableConfigTypeBase> (db: string, table: string): (<const I extends ExtraTableConfigIndexBase<T>> (indexes: I) => ExtraTableConfig<T, I>) {
-  return <const I extends ExtraTableConfigIndexBase<T>> (indexes: I): ExtraTableConfig<T, I> => {
-    return { db, table, type: null as unknown as T, indexes }
+export function table<T extends ExtraTableConfigTypeBase>(_options?: unknown) {
+  /**
+   * Describe the indexes in this table. If you have no indexes, just pass `{}`.
+   * 
+   * Otherwise, pass a map of index name to config object, like so:
+   * ```ts
+   * simple: {}, // (on 'simple' field)
+   * multi: { multi: true }, // (on 'multi' field)
+   * compound: { compound: ['last_name', 'first_name'] }, // named 'compound'
+   * custom: { custom: function (row) { return row('hobbies').add(row('sports')) } } // named 'custom'
+   * ```
+   * 
+   * For more info, refer to the wiki page on [Secondary Indexes](https://github.com/dawnniie/rethinkdb-ts-extra/wiki/Secondary-Indexes).
+   */
+  return <const I extends ExtraTableConfigIndexBase<T>>(indexes: I) => {
+    const configuration: ExtraTableConfig<T, I> = { type: null as unknown as T, indexes }
+    return configuration
   }
 }
 
-type Re<Configs extends { [name: string]: ExtraTableConfig<any, any> }> = typeof r
-  & { [tableName in keyof Configs]: RTableExtra<Configs[tableName]> }
-  & {
-    /** Wrap around an index name to perform orderBy in reverse */
-    desc<T>(index: T): T,
-    /** A typescript-friendly way of clearing a field using `re.literal()` */
-    empty(): Empty
+/**
+ * Initialise `rethinkdb-ts-extra` and attach types to the `r` object before returning it.
+ * 
+ * Please refer to [the wiki (Getting Started)](https://github.com/dawnniie/rethinkdb-ts-extra/wiki/Getting-Started) for guidance on how to use this function, as it goes into more detail for the various options.
+ * 
+ * The first option should always be your db & table configurations, which is a two-layer object mapping table names to the `table` function, like so:
+ * ```ts
+ * extra({
+ *   db_name: {
+ *     table_name: table<...>()({ ... }),
+ *     ...
+ *   },
+ *   ...
+ * })
+ * ```
+ */
+async function extra<const Configs extends RDatabaseExtraConfigs, const DefaultDB extends string = 'test'>(configs: Configs): Promise<RExtra<Configs, DefaultDB>>
+async function extra<const Configs extends RDatabaseExtraConfigs, const DefaultDB extends string = 'test'>(configs: Configs, defaultDb: DefaultDB): Promise<RExtra<Configs, DefaultDB>>
+async function extra<const Configs extends RDatabaseExtraConfigs, const DefaultDB extends string = 'test'>(configs: Configs, connect: true | ConnectOptions<DefaultDB>, sync?: SyncOptions): Promise<RExtra<Configs, DefaultDB>>
+async function extra<const Configs extends RDatabaseExtraConfigs, const DefaultDB extends string = 'test'>(configs: Configs, defaultDbOrConnect?: DefaultDB | true | ConnectOptions<DefaultDB>, syncOptions?: true | SyncOptions) {
+  // @ts-expect-error this override is allowed, as long as we keep it typed correctly
+  const r = _r as RExtra<Configs, DefaultDB>
+
+  r.extra = {
+    connect: async (options?: ConnectOptions<DefaultDB>) => await connect(r, options),
+    sync: async (options?: SyncOptions) => await sync(r as unknown as R, configs, options)
   }
 
-interface AttachConfigurationsReturnType<Configs extends { [name: string]: ExtraTableConfig<any, any> }> { r: Re<Configs>, sync: ReturnType<typeof createSyncer> }
+  // @ts-expect-error type otherwise defined
+  r.$ = (dbName: string, tableName?: string) => tableName ? r.db(dbName).table(tableName) : r.table(dbName)
 
-/**
- * Attaches provided table configurations (and other utilies) to the `r` object with detailed extra data and index typings. Also provides a `sync` function for in-place database structure updates (sort of like automated structure (not data!) migrations).
- *
- * Each table config is attached to `r` with its name (the key in the object), so that the table can be accessed by shorthand like so:
- *
- * ```ts
- * const { r } = attachConfigurations({ myTableName: configure<{ id: string }>('mydb', 'mytable')({}) })
- *
- * // instead of
- * r.db('mydb').table('mytable') // -> RTable<any>
- * // you can do
- * r.myTableName // -> RTableExtra<{ id: string , ...>
- * ```
- *
- * The type `RTableExtra` is an extension of `RTable`. While `RTable` already supports providing types, this must be done manually with each call. The shorthand pre-populates this type information, as well as extra information for secondary indexes. Some key `RTable` methods (currently `get`, `getAll`, `between` and `insert`) are re-typed in `RTableExtra` to take advantage of this extra type information. When you run an operation with an index, the configuration for that index and the table data will be used to typecheck the data you are passing.
- *
- * Even though these shorthands are attached in-place to the `r` object exported by `rethinkdb-ts`, you likely want to export and reference the `r` object where it is returned by `attachConfigurations`, because this is the only way to get the correct typings.
- *
- * If you're using custom indexes, you may find that having the `configure` calls in the arguments for the `attachConfigurations` call breaks the inference of the custom expression argument type. I can't figure out how to fix this yet, and my only solution has been to move the configure call away and pass the output in. If you have any ideas for this, please let me know.
- */
-export function attachConfigurations<const Configs extends { [name: string]: ExtraTableConfig<any, any> }> (configs: Configs): AttachConfigurationsReturnType<Configs> {
-  // @ts-expect-error this override is defined below
-  for (const [name, config] of Object.entries(configs)) r[name] = r.db(config.db).table(config.table)
-  // @ts-expect-error this override is defined below
-  r.empty = () => r.literal()
+  r[Symbol.asyncDispose] = async () => await r.getPoolMaster()?.drain()
+  
+  if (defaultDbOrConnect && typeof defaultDbOrConnect !== 'string') {
+    // we should connect for the user
+    await r.extra.connect(defaultDbOrConnect === true ? {} : defaultDbOrConnect)
+    if (syncOptions) await r.extra.sync(syncOptions === true ? {} : syncOptions)
+  } else {
+    // the user is managing connecting themselves
+    // we should do some level of enforcing/notification for inconsistent default db settings, in case they mess it up
+    const defaultDb = defaultDbOrConnect || 'test'
+    function checkConnections() {
+      const connections = r.getPoolMaster()?.getPools().map(p => p.getConnections()).flat()
+      // @ts-expect-error allow us to reach in to this private property: https://github.com/rethinkdb/rethinkdb-ts/blob/main/src/connection/connection.ts#L44
+      if (connections?.find(c => c.db !== defaultDb)) console.warn('warning: default db option set, but connections detected with different option - this will cause unexpected behaviour when using queries that assume a default database')
+      // keep trying until we actually find a connection to check
+      if (!connections?.length) setTimeout(() => { checkConnections(); }, 10 * 1000)
+    }
 
-  return { r: r as Re<Configs>, sync: createSyncer(r, configs) }
+    setTimeout(() => { checkConnections(); }, 10 * 1000)
+  }
+
+  return r
 }
 
-export * from 'rethinkdb-ts'
+export { extra }